Resolves #52 (#54)
Found code online that creates the smallest circle given some points. Included and adapted that code for our use.

The problem of determining if three points can fit inside a circle with a given radius turned out to be harder than expected. We need to be able to determine that for a few LICs, e.g. 1 and 8.